Refactoring pyapnea to help its use. Improving tutorial and documentation. V0.0.5a
