+++ docs/index.md
@@ -1,7 +1,10 @@
-# Welcome to pyapnea |version|
+# Welcome to pyapnea 0.0.5a
 
 The Python package related to apnea data analysis.
 
-This version contains functions to read session files from OSCAR software (https://www.sleepfiles.com/OSCAR/)
+This version contains functions to read session files from OSCAR software (https://www.sleepfiles.com/OSCAR/) and a pytorch dataset to read Oscar data
+
+For a more general view of functionalities of this library, a tutorial can help you.
 
 See API for more details about functions in this library.
+
