+++ pyapnea/oscar/oscar_constants.py
